requirements.txt
streamlit
ultralytics
opencv-python-headless
torch
Pillow
numpy